requirements.txt
streamlit>=1.36.0
supabase>=2.4.0
python-dotenv>=1.0.1
pandas>=2.2.2
